Fixed Door Bug Refactored DungeonManager Classes Fixed Player and Door interactions Fixed player spawning Added Enums for DoorDirection and RoomType Refactored DungeonManager Class Visibility to be as strict as possible

+++ src/model/GameObject.java
@@ -4,6 +4,6 @@ public class GameObject {
     public enum  FacingDirection {
         UP, DOWN, LEFT, RIGHT
     }
-    public int myX, myY;
+    public double myX, myY;
     public int mySpeed;
 }
